Add full dependencies release and OpenCV caching

+++ build_scripts/components/opencv.sh
@@ -1,5 +1,6 @@
 #!/bin/bash -e
 
+OPENCV_VERSION="4.1.0"
 DOCKER="docker"
 
 if ! ${DOCKER} ps >/dev/null 2>&1; then
@@ -11,9 +12,17 @@ if ! ${DOCKER} ps >/dev/null; then
         exit 1
 fi
 
-${DOCKER} build -t opsi-opencv opencv-build
-${DOCKER} run --rm --privileged \
-    --volume "$(pwd)":/packages \
-    --name "opencv-build_work" \
-    opsi-opencv \
-    bash -e -o pipefail -c "cd /packages/opencv-build/; ./build.sh"
+if [ ! -e "cache/opencv-python_${OPENCV_VERSION}_armhf.deb" ]; then
+    ${DOCKER} build -t opsi-opencv opencv-build
+    ${DOCKER} run --rm --privileged \
+        --volume "$(pwd)":/packages \
+        --name "opencv-build_work" \
+        opsi-opencv \
+        bash -e -o pipefail -c "cd /packages/opencv-build/; OPENCV_VERSION=${OPENCV_VERSION} ./build.sh"
+    rm -rf cache/*
+    cp "packages/opencv-python_${OPENCV_VERSION}_armhf.deb" "cache/opencv-python_${OPENCV_VERSION}_armhf.deb"
+    cp "packages/opencv-libs_${OPENCV_VERSION}_armhf.deb" "cache/opencv-libs_${OPENCV_VERSION}_armhf.deb"
+else
+    cp "cache/opencv-python_${OPENCV_VERSION}_armhf.deb" "packages/opencv-python_${OPENCV_VERSION}_armhf.deb"
+    cp "cache/opencv-libs_${OPENCV_VERSION}_armhf.deb" "packages/opencv-libs_${OPENCV_VERSION}_armhf.deb"
+fi
